- Arreglamos las ventas como jInternalFrames separados - Se creo RegistrarRestaurante y RegistrarProducto.
