+++ Tarea1/nbproject/private/private.properties
@@ -3,4 +3,4 @@ do.depend=false
 do.jar=true
 javac.debug=true
 javadoc.preview=true
-user.properties.file=C:\\Users\\LEO\\AppData\\Roaming\\NetBeans\\8.0.2\\build.properties
+user.properties.file=/home/natalia/.netbeans/8.0.2/build.properties
