+++ Tarea1/src/Presentacion/InfoCliente.java
@@ -1,0 +1,192 @@
+/*
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
+package Presentacion;
+
+import Logica.DataUsuario;
+import Logica.IControladorCategoria;
+import Logica.IControladorPedido;
+import Logica.IControladorProducto;
+import Logica.IControladorUsuario;
+
+/**
+ *
+ * @author natalia
+ */
+public class InfoCliente extends javax.swing.JInternalFrame {
+    private IControladorUsuario ICU;
+    private IControladorProducto ICP;
+    private IControladorPedido ICPE;
+    private IControladorCategoria ICC;
+    private String d,m,a;
+    private int D,M,A;
+
+    /**
+     * Creates new form InfoCliente
+     */
+    public InfoCliente() {
+        initComponents();
+    }
+
+    /**
+     * This method is called from within the constructor to initialize the form.
+     * WARNING: Do NOT modify this code. The content of this method is always
+     * regenerated by the Form Editor.
+     */
+    @SuppressWarnings("unchecked")
+    // <editor-fold defaultstate="collapsed" desc="Generated Code">//GEN-BEGIN:initComponents
+    private void initComponents() {
+
+        jScrollPane3 = new javax.swing.JScrollPane();
+        listadeclientes = new javax.swing.JTable();
+        MostrarInfo = new javax.swing.JButton();
+        jButton5 = new javax.swing.JButton();
+        ingresonick = new javax.swing.JTextField();
+        nick = new javax.swing.JTextField();
+        nombre = new javax.swing.JTextField();
+        apellido = new javax.swing.JTextField();
+
+        setClosable(true);
+        setIconifiable(true);
+
+        listadeclientes.setAutoCreateRowSorter(true);
+        listadeclientes.setModel(new javax.swing.table.DefaultTableModel(
+            new Object [][] {
+                {"nacho", "sdfsdfsd"},
+                {"leo", "sdfsdfs"},
+                {"nati", "fsdfsdf"},
+                {null, null},
+                {null, null}
+            },
+            new String [] {
+                "nickname", "email"
+            }
+        ) {
+            Class[] types = new Class [] {
+                java.lang.String.class, java.lang.String.class
+            };
+            boolean[] canEdit = new boolean [] {
+                false, false
+            };
+
+            public Class getColumnClass(int columnIndex) {
+                return types [columnIndex];
+            }
+
+            public boolean isCellEditable(int rowIndex, int columnIndex) {
+                return canEdit [columnIndex];
+            }
+        });
+        listadeclientes.setCursor(new java.awt.Cursor(java.awt.Cursor.DEFAULT_CURSOR));
+        jScrollPane3.setViewportView(listadeclientes);
+
+        MostrarInfo.setText("Mostrar Informacion");
+        MostrarInfo.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                MostrarInfoActionPerformed(evt);
+            }
+        });
+
+        jButton5.setText("jButton5");
+        jButton5.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                jButton5ActionPerformed(evt);
+            }
+        });
+
+        nick.setText("jTextField2");
+
+        nombre.setText("jTextField3");
+
+        apellido.setText("jTextField4");
+
+        javax.swing.GroupLayout layout = new javax.swing.GroupLayout(getContentPane());
+        getContentPane().setLayout(layout);
+        layout.setHorizontalGroup(
+            layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(layout.createSequentialGroup()
+                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                    .addGroup(layout.createSequentialGroup()
+                        .addContainerGap()
+                        .addComponent(jScrollPane3, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
+                    .addGroup(layout.createSequentialGroup()
+                        .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                            .addGroup(layout.createSequentialGroup()
+                                .addGap(142, 142, 142)
+                                .addComponent(MostrarInfo, javax.swing.GroupLayout.PREFERRED_SIZE, 181, javax.swing.GroupLayout.PREFERRED_SIZE))
+                            .addGroup(layout.createSequentialGroup()
+                                .addGap(55, 55, 55)
+                                .addComponent(ingresonick, javax.swing.GroupLayout.PREFERRED_SIZE, 100, javax.swing.GroupLayout.PREFERRED_SIZE)
+                                .addGap(41, 41, 41)
+                                .addComponent(jButton5)))
+                        .addGap(24, 24, 24)
+                        .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.TRAILING)
+                            .addComponent(nombre, javax.swing.GroupLayout.DEFAULT_SIZE, 144, Short.MAX_VALUE)
+                            .addComponent(apellido)
+                            .addComponent(nick))))
+                .addContainerGap(40, Short.MAX_VALUE))
+        );
+        layout.setVerticalGroup(
+            layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(layout.createSequentialGroup()
+                .addContainerGap()
+                .addComponent(jScrollPane3, javax.swing.GroupLayout.PREFERRED_SIZE, 149, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                    .addGroup(layout.createSequentialGroup()
+                        .addGap(18, 18, 18)
+                        .addComponent(MostrarInfo))
+                    .addGroup(layout.createSequentialGroup()
+                        .addGap(45, 45, 45)
+                        .addComponent(nick, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                        .addComponent(nombre, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                        .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                            .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                                .addComponent(ingresonick, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                                .addComponent(jButton5))
+                            .addComponent(apellido, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))))
+                .addContainerGap(56, Short.MAX_VALUE))
+        );
+
+        pack();
+    }// </editor-fold>//GEN-END:initComponents
+
+    private void MostrarInfoActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_MostrarInfoActionPerformed
+        DataUsuario du=ICU.Caso_Ver_Cliente(this.ingresonick.getText(), "");
+        this.nick.setText(du.getnickname());
+        this.nombre.setText(du.getnombre());
+        this.apellido.setText(du.getapellido());
+        /*this.VentRegCli.setVisible(true);
+        this.SeleccionarImagen.setVisible(false);
+        this.Registrar.setVisible(false);
+        this.txtApellido.setEditable(false);
+        this.txtCorreoElectronico.setEditable(false);
+        this.txtDireccion.setEditable(false);
+        this.txtNickname.setEditable(false);
+        this.txtNombre.setEditable(false);*/
+    }//GEN-LAST:event_MostrarInfoActionPerformed
+
+    private void jButton5ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jButton5ActionPerformed
+        this.listadeclientes.setValueAt("aaaa", 0, 0);
+
+        int fila = this.listadeclientes.getSelectedRow();
+        int columna = this.listadeclientes.getSelectedColumn();
+        String nickname=(String)this.listadeclientes.getValueAt(fila, columna);
+        this.ingresonick.setText(nickname);
+    }//GEN-LAST:event_jButton5ActionPerformed
+
+
+    // Variables declaration - do not modify//GEN-BEGIN:variables
+    private javax.swing.JButton MostrarInfo;
+    private javax.swing.JTextField apellido;
+    private javax.swing.JTextField ingresonick;
+    private javax.swing.JButton jButton5;
+    private javax.swing.JScrollPane jScrollPane3;
+    public static javax.swing.JTable listadeclientes;
+    private javax.swing.JTextField nick;
+    private javax.swing.JTextField nombre;
+    // End of variables declaration//GEN-END:variables
+}
